+++ contextualized_topic_models/models/ctm.py
@@ -23,8 +23,8 @@ class CTM:
     """Class to train the contextualized topic model. This is the more general class that we are keeping to
     avoid braking code, user should use the two subclasses ZeroShotTM and CombinedTm to do topic modeling.
 
-        :param input_size: int, dimension of input
-        :param bert_input_size: int, dimension of input that comes from BERT embeddings
+        :param bow_size: int, dimension of input
+        :param contextual_size: int, dimension of input that comes from BERT embeddings
         :param inference_type: string, you can choose between the contextual model and the combined model
         :param n_components: int, number of topic components, (default 10)
         :param model_type: string, 'prodLDA' or 'LDA' (default 'prodLDA')
@@ -41,20 +41,18 @@ class CTM:
         :param num_data_loader_workers: int, number of data loader workers (default cpu_count). set it to 0 if you are using Windows
     """
 
-    def __init__(self, input_size, bert_input_size, inference_type="combined", n_components=10, model_type='prodLDA',
+    def __init__(self, bow_size, contextual_size, inference_type="combined", n_components=10, model_type='prodLDA',
                  hidden_sizes=(100, 100), activation='softplus', dropout=0.2,
                  learn_priors=True, batch_size=64, lr=2e-3, momentum=0.99,
                  solver='adam', num_epochs=100, reduce_on_plateau=False, num_data_loader_workers=mp.cpu_count()):
-        warnings.simplefilter('always', DeprecationWarning)
+
 
         if self.__class__.__name__ == "CTM":
-            warnings.warn(
-                "Direct call to CTM is deprecated and will be removed in version 2, use CombinedTM or ZeroShotTM",
-                DeprecationWarning)
+            raise Exception("You cannot call this class. Use ZeroShotTM or CombinedTM")
 
-        assert isinstance(input_size, int) and input_size > 0, \
+        assert isinstance(bow_size, int) and bow_size > 0, \
             "input_size must by type int > 0."
-        assert isinstance(n_components, int) and input_size > 0, \
+        assert isinstance(n_components, int) and bow_size > 0, \
             "n_components must by type int > 0."
         assert model_type in ['LDA', 'prodLDA'], \
             "model must be 'LDA' or 'prodLDA'."
@@ -75,7 +73,7 @@ class CTM:
         assert isinstance(num_data_loader_workers, int) and num_data_loader_workers >= 0, \
             "num_data_loader_workers must by type int >= 0. set 0 if you are using windows"
 
-        self.input_size = input_size
+        self.bow_size = bow_size
         self.n_components = n_components
         self.model_type = model_type
         self.hidden_sizes = hidden_sizes
@@ -84,7 +82,7 @@ class CTM:
         self.learn_priors = learn_priors
         self.batch_size = batch_size
         self.lr = lr
-        self.bert_size = bert_input_size
+        self.contextual_size = contextual_size
         self.momentum = momentum
         self.solver = solver
         self.num_epochs = num_epochs
@@ -92,7 +90,7 @@ class CTM:
         self.num_data_loader_workers = num_data_loader_workers
 
         self.model = DecoderNetwork(
-            input_size, self.bert_size, inference_type, n_components, model_type, hidden_sizes, activation,
+            bow_size, self.contextual_size, inference_type, n_components, model_type, hidden_sizes, activation,
             dropout, learn_priors)
         self.early_stopping = None
 
@@ -315,9 +313,6 @@ class CTM:
         :param dataset: a PyTorch Dataset containing the documents
         :param n_samples: the number of sample to collect to estimate the final distribution (the more the better).
         """
-        warnings.warn("Call to `get_thetas` is deprecated and will be removed in version 2, "
-                      "use `get_doc_topic_distribution` instead",
-                      DeprecationWarning)
         return self.get_doc_topic_distribution(dataset, n_samples=n_samples)
 
     def get_doc_topic_distribution(self, dataset, n_samples=20):
@@ -404,7 +399,7 @@ class CTM:
 
         :param k: int, number of words to return per topic, default 10.
         """
-        assert k <= self.input_size, "k must be <= input size."
+        assert k <= self.bow_size, "k must be <= input size."
         component_dists = self.best_components
         topics = defaultdict(list)
         for i in range(self.n_components):
@@ -420,7 +415,7 @@ class CTM:
 
         :param k: (int) number of words to return per topic, default 10.
         """
-        assert k <= self.input_size, "k must be <= input size."
+        assert k <= self.bow_size, "k must be <= input size."
         # TODO: collapse this method with the one that just returns the topics
         component_dists = self.best_components
         topics = []
@@ -555,8 +550,8 @@ class ZeroShotTM(CTM):
     """
     ZeroShotTM, as described in https://arxiv.org/pdf/2004.07737v1.pdf
 
-    :param input_size: int, dimension of input
-    :param bert_input_size: int, dimension of input that comes from BERT embeddings
+    :param bow_size: int, dimension of input
+    :param contextual_size: int, dimension of input that comes from BERT embeddings
     :param n_components: int, number of topic components, (default 10)
     :param model_type: string, 'prodLDA' or 'LDA' (default 'prodLDA')
     :param hidden_sizes: tuple, length = n_layers, (default (100, 100))
@@ -570,15 +565,14 @@ class ZeroShotTM(CTM):
     :param num_epochs: int, number of epochs to train for, (default 100)
     :param reduce_on_plateau: bool, reduce learning rate by 10x on plateau of 10 epochs (default False)
     :param num_data_loader_workers: int, number of data loader workers (default cpu_count). set it to 0 if you are using Windows
-
     """
 
-    def __init__(self, input_size, bert_input_size, n_components=10, model_type='prodLDA',
+    def __init__(self, bow_size, contextual_size, n_components=10, model_type='prodLDA',
                  hidden_sizes=(100, 100), activation='softplus', dropout=0.2,
                  learn_priors=True, batch_size=64, lr=2e-3, momentum=0.99,
                  solver='adam', num_epochs=100, reduce_on_plateau=False, num_data_loader_workers=mp.cpu_count()):
         inference_type = "zeroshot"
-        super().__init__(input_size, bert_input_size, inference_type, n_components, model_type,
+        super().__init__(bow_size, contextual_size, inference_type, n_components, model_type,
                          hidden_sizes, activation, dropout,
                          learn_priors, batch_size, lr, momentum,
                          solver, num_epochs, reduce_on_plateau, num_data_loader_workers)
@@ -588,8 +582,8 @@ class CombinedTM(CTM):
     """
     CombinedTM, as described in https://arxiv.org/pdf/2004.03974.pdf
 
-    :param input_size: int, dimension of input
-    :param bert_input_size: int, dimension of input that comes from BERT embeddings
+    :param bow_size: int, dimension of input
+    :param contextual_size: int, dimension of input that comes from BERT embeddings
     :param n_components: int, number of topic components, (default 10)
     :param model_type: string, 'prodLDA' or 'LDA' (default 'prodLDA')
     :param hidden_sizes: tuple, length = n_layers, (default (100, 100))
@@ -605,12 +599,12 @@ class CombinedTM(CTM):
     :param num_data_loader_workers: int, number of data loader workers (default cpu_count). set it to 0 if you are using Windows
     """
 
-    def __init__(self, input_size, bert_input_size, n_components=10, model_type='prodLDA',
+    def __init__(self, bow_size, contextual_size, n_components=10, model_type='prodLDA',
                  hidden_sizes=(100, 100), activation='softplus', dropout=0.2,
                  learn_priors=True, batch_size=64, lr=2e-3, momentum=0.99,
                  solver='adam', num_epochs=100, reduce_on_plateau=False, num_data_loader_workers=mp.cpu_count()):
         inference_type = "combined"
-        super().__init__(input_size, bert_input_size, inference_type, n_components, model_type,
+        super().__init__(bow_size, contextual_size, inference_type, n_components, model_type,
                          hidden_sizes, activation, dropout,
                          learn_priors, batch_size, lr, momentum,
                          solver, num_epochs, reduce_on_plateau, num_data_loader_workers)